requirements.txt
Django==1.8.13
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-pagedown==0.1.1
Markdown==2.6.6
markdown2==2.3.1
pytz==2016.6.1
